Moved even more common stuff into the template
